pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^5.0"
python = "^3.10"
requests = "^2.32.5"
# MCP and AI Agent Dependencies
openai = "^1.12.0"
anthropic = "^0.18.0"
mcp = "^1.16.0"
langchain = "^0.1.0"
langchain-community = "^0.0.20"
langchain-google-community = "^0.0.13"
# Google Calendar Integration
google-auth = "^2.25.2"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.110.0"
# Enhanced Financial Features
pandas = "^2.1.4"
numpy = "^1.26.3"
# Database and Security
psycopg2-binary = "^2.9.9"
cryptography = "^42.0.4"
# API and HTTP
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
httpx = "^0.26.0"
# JSON-RPC for MCP
jsonrpc-base = "^1.0.4"
# Environment and Configuration
python-dotenv = "^1.0.0"
# Testing and Development
pytest = "^8.0.0"
pytest-asyncio = "^0.23.2"
pytest-django = "^4.7.0"
[tool.poetry.dev-dependencies]

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"